update split merge lod tensor, test=develop
