+++ lite/backends/arm/math/CMakeLists.txt
@@ -119,6 +119,7 @@ if (NOT HAS_ARM_MATH_LIB_DIR)
       stack.cc
       affine_channel.cc
       anchor_generator.cc
+      split_merge_lod_tenosr.cc
       DEPS ${lite_kernel_deps} context tensor)
 endif()
